requirements.txt
streamlit
python-dotenv

openai
langchain
langchain-core
langchain-openai
langchain-chroma
chromadb
langchain-text-splitters

pytest